+++ src/other/broadcasting.jl
@@ -125,9 +125,12 @@ Base.dotview(df::SubDataFrame, ::typeof(!), idxs) =
 if isdefined(Base, :dotgetproperty)
     Base.dotgetproperty(df::DataFrame, col::SymbolOrString) =
         LazyNewColDataFrame(df, Symbol(col))
-    Base.dotgetproperty(df::SubDataFrame, col::SymbolOrString) =
-        throw(ArgumentError("broadcasting getproperty is not allowed for SubDataFrame since " *
-                            "Julia 1.7. use `df[:, $col] .= ... instead"))
+
+    function Base.dotgetproperty(df::SubDataFrame, col::SymbolOrString)
+        Base.depwarn("broadcasting getproperty is deprecated for SubDataFrame " *
+                    "since Julia 1.7. Use `df[:, $col] .= ... instead", :dotgetproperty)
+        return getproperty(df, col)
+    end
 end
 
 function Base.copyto!(lazydf::LazyNewColDataFrame, bc::Base.Broadcast.Broadcasted{T}) where T